small patches to implementations. Nothing new
